requirements.txt
Django==3.2.16
djangorestframework==3.14.0
psycopg2==2.8.6
aiohttp==3.8
openpyxl==3.1.1
pydantic==1.10
drf-yasg==1.21.4
uvicorn==0.19.0
requests~=2.28.2
whitenoise==6.3.0